fix(lineitemstate): remove invalid logger defaults
the logger default added was `process.stderr` and it's expected to be called like

`process.stderr()` this will throw an error if called because there is no method like that

what

is present is `process.stderr.write()` because it's a stream. Two solutions is to bind or use

`console.log`.

 I went with `console.log` ;)