1 implement  main module (#7)
* * feat(lineitemstateupdate): add lineitemstate  submodule

* feat(lineitemstateupdate): add initial class module

* feat(lineitemstate): add parse and process method

* fix(CLI): chmod CLI file to be executable

* Add returninfo submodule (#3)

* feat(lineitemstateupdate): add initial class module

* feat(lineitemstate): add parse and process method

* feat(add-returninfo): implement submodule

* refactor(return-info): make code better

* WIP

* refactor reduce some more

* flatten the if else statements

* Implement cli (#6)

* feat(get-api-credentials): add method

* feat(CLI): add basic implementation

* feat(CLI): add returnInfo module to CLI

* chore(package.json): improve npm coverage script

* style(variable): do not declare the parsed result as an additional variable

* fix(returninfo): remove invalid logger defaults

Logger default added was `process.stderr` and it's expected to be called like

`process.stderr()`

this will throw an error if called because there is no method like that

what is

present is `process.stderr.write()` because it's a stream. Two solutions is to bind or use

`console.log`.

I went with `console.log` ;)

* fix(lineitemstate): remove invalid logger defaults

the logger default added was `process.stderr` and it's expected to be called like

`process.stderr()` this will throw an error if called because there is no method like that

what

is present is `process.stderr.write()` because it's a stream. Two solutions is to bind or use

`console.log`.

 I went with `console.log` ;)

* chore(logging): add a start conversion info

* chore(CLI): update log file name

* refactor(userAgent): use sphere-node-utils module to build userAgent

* style(src): remove unused eslint disable command

* docs(readme): update readme

* chore(package.json): add entry file and bin field

* test(cli): make tests 100% 🌞